backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
alembic>=1.13.1
psycopg2-binary>=2.9.10
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.10.0
python-multipart>=0.0.9
websockets>=12.0
httpx>=0.27.0
python-dotenv>=1.0.1
